Move Sink and Source embassy implementations to own module
