+++ heimlig/src/client/api.rs
@@ -44,93 +44,11 @@ impl<
 mod test {
     use crate::client::api::Api;
     use crate::common::jobs::{Request, Response};
-    use core::cell::RefCell;
-    use core::pin::Pin;
-    use core::task::Context;
-    use core::task::Poll;
-    use embassy_sync::blocking_mutex;
+    use crate::integration::embassy::{RequestQueueSink, ResponseQueueSource};
     use embassy_sync::blocking_mutex::raw::NoopRawMutex;
-    use embassy_sync::blocking_mutex::raw::RawMutex;
-    use embassy_sync::waitqueue::WakerRegistration;
-    use heapless::spsc::{Consumer, Producer, Queue};
+    use heapless::spsc::Queue;
 
     const QUEUE_SIZE: usize = 8;
-
-    struct RequestQueueSink<'ch, 'data, M: RawMutex + Unpin> {
-        producer: Producer<'ch, Request<'data>, QUEUE_SIZE>,
-        receiver_waker: blocking_mutex::Mutex<M, RefCell<WakerRegistration>>,
-        sender_waker: blocking_mutex::Mutex<M, RefCell<WakerRegistration>>,
-    }
-
-    impl<'data, M: RawMutex + Unpin> futures::Sink<Request<'data>> for RequestQueueSink<'_, 'data, M> {
-        type Error = ();
-
-        fn poll_ready(
-            self: Pin<&mut Self>,
-            _cx: &mut Context<'_>,
-        ) -> Poll<Result<(), Self::Error>> {
-            if self.producer.ready() {
-                Poll::Ready(Ok(()))
-            } else {
-                self.sender_waker.lock(|w| w.borrow_mut().wake());
-                Poll::Pending
-            }
-        }
-
-        fn start_send(self: Pin<&mut Self>, request: Request<'data>) -> Result<(), Self::Error> {
-            self.receiver_waker.lock(|w| w.borrow_mut().wake());
-            let _request = self.get_mut().producer.enqueue(request);
-            Ok(())
-        }
-
-        fn poll_flush(
-            self: Pin<&mut Self>,
-            _cx: &mut Context<'_>,
-        ) -> Poll<Result<(), Self::Error>> {
-            Poll::Ready(Ok(()))
-        }
-
-        fn poll_close(
-            self: Pin<&mut Self>,
-            _cx: &mut Context<'_>,
-        ) -> Poll<Result<(), Self::Error>> {
-            Poll::Ready(Ok(()))
-        }
-    }
-
-    impl<'ch, 'data, M: RawMutex + Unpin> RequestQueueSink<'ch, 'data, M> {
-        pub fn new(requests: Producer<'ch, Request<'data>, QUEUE_SIZE>) -> Self {
-            RequestQueueSink {
-                producer: requests,
-                receiver_waker: blocking_mutex::Mutex::new(RefCell::new(WakerRegistration::new())),
-                sender_waker: blocking_mutex::Mutex::new(RefCell::new(WakerRegistration::new())),
-            }
-        }
-    }
-
-    struct ResponseQueueSource<'ch, 'data, M: RawMutex + Unpin> {
-        consumer: Consumer<'ch, Response<'data>, QUEUE_SIZE>,
-        senders_waker: blocking_mutex::Mutex<M, RefCell<WakerRegistration>>,
-    }
-
-    impl<'data, M: RawMutex + Unpin> futures::Stream for ResponseQueueSource<'_, 'data, M> {
-        type Item = Response<'data>;
-
-        fn poll_next(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-            self.senders_waker.lock(|w| w.borrow_mut().wake());
-            // No need to return pending and wake a receiver waker as dequeue() always returns directly
-            Poll::Ready(self.get_mut().consumer.dequeue())
-        }
-    }
-
-    impl<'ch, 'data, M: RawMutex + Unpin> ResponseQueueSource<'ch, 'data, M> {
-        pub fn new(responses: Consumer<'ch, Response<'data>, QUEUE_SIZE>) -> Self {
-            ResponseQueueSource {
-                consumer: responses,
-                senders_waker: blocking_mutex::Mutex::new(RefCell::new(WakerRegistration::new())),
-            }
-        }
-    }
 
     #[async_std::test]
     async fn send_request() {
@@ -144,8 +62,10 @@ mod test {
         let (requests_tx, mut requests_rx) = requests.split();
         let (mut responses_tx, responses_rx) = responses.split();
 
-        let requests: RequestQueueSink<NoopRawMutex> = RequestQueueSink::new(requests_tx);
-        let responses: ResponseQueueSource<NoopRawMutex> = ResponseQueueSource::new(responses_rx);
+        let requests: RequestQueueSink<NoopRawMutex, QUEUE_SIZE> =
+            RequestQueueSink::new(requests_tx);
+        let responses: ResponseQueueSource<NoopRawMutex, QUEUE_SIZE> =
+            ResponseQueueSource::new(responses_rx);
 
         let mut api = Api::new(requests, responses);
