+++ heimlig/src/lib.rs
@@ -5,3 +5,4 @@ pub mod common;
 pub mod config;
 pub mod crypto;
 pub mod hsm;
+pub mod integration;
